Use Statistics::Basic instead of hand-rolled code.
One of the most important habits to get into as a Perl programmer is the habit
of asking yourself "is there a CPAN module that could help me out here?" Hint:
yes, there is. In this case, Statistics::Basic allows you to (among other
things) calculate means and standard deviations without writing
potentially-buggy code yourself.

We also move the declarations of $(control|sample)(Mean|SD) to their point of
first use, limiting their visibility to the block in which they're actually
used.